+++ internal/client/client.go
@@ -30,6 +30,7 @@ type ZNode struct {
 	Path string
 	Stat *zk.Stat
 	Data []byte
+	ACL  []zk.ACL
 }
 
 // Re-exporting errors from ZK library for better encapsulation.
@@ -64,15 +65,32 @@ const (
 	// DefaultZooKeeperSessionSec is the default amount of seconds configured for the
 	// Client timeout session, in case EnvZooKeeperSessionSec is not set.
 	DefaultZooKeeperSessionSec = 30
+
+	// Environment variables to provide digest auth credentials.
+	EnvZooKeeperUsername = "ZOOKEEPER_USERNAME"
+	EnvZooKeeperPassword = "ZOOKEEPER_PASSWORD"
 )
 
 // NewClient constructs a new Client instance.
-func NewClient(servers string, sessionTimeoutSec int) (*Client, error) {
+func NewClient(servers string, sessionTimeoutSec int, username string, password string) (*Client, error) {
 	serversSplit := strings.Split(servers, serversStringSeparator)
 
 	conn, _, err := zk.Connect(zk.FormatServers(serversSplit), time.Duration(sessionTimeoutSec)*time.Second)
 	if err != nil {
 		return nil, fmt.Errorf("unable to connect to ZooKeeper: %w", err)
+	}
+
+	if (username == "") != (password == "") {
+		return nil, fmt.Errorf("both username and password must be specified together")
+	}
+
+	if username != "" {
+		auth := "digest"
+		credentials := fmt.Sprintf("%s:%s", username, password)
+		err = conn.AddAuth(auth, []byte(credentials))
+		if err != nil {
+			return nil, fmt.Errorf("unable to add digest auth: %w", err)
+		}
 	}
 
 	return &Client{
@@ -99,16 +117,16 @@ func NewClientFromEnv() (*Client, error) {
 		return nil, fmt.Errorf("failed to convert '%s' to integer: %w", zkSession, err)
 	}
 
-	return NewClient(zkServers, zkSessionInt)
+	zkUsername, _ := os.LookupEnv(EnvZooKeeperUsername)
+	zkPassword, _ := os.LookupEnv(EnvZooKeeperPassword)
+
+	return NewClient(zkServers, zkSessionInt, zkUsername, zkPassword)
 }
 
 // Create a ZNode at the given path.
 //
 // Note that any necessary ZNode parents will be created if absent.
-func (c *Client) Create(path string, data []byte) (*ZNode, error) {
-	// TODO Make ACL configurable
-	acl := zk.WorldACL(zk.PermRead | zk.PermWrite | zk.PermCreate | zk.PermDelete)
-
+func (c *Client) Create(path string, data []byte, acl []zk.ACL) (*ZNode, error) {
 	if path[len(path)-1] == zNodePathSeparator {
 		return nil, fmt.Errorf("non-sequential ZNode cannot have path '%s' because it ends in '%c'", path, zNodePathSeparator)
 	}
@@ -130,10 +148,7 @@ func (c *Client) Create(path string, data []byte) (*ZNode, error) {
 //   - created znode path -> `/this/is/a/path/0000000001`
 //
 // Note also that any necessary ZNode parents will be created if absent.
-func (c *Client) CreateSequential(path string, data []byte) (*ZNode, error) {
-	// TODO Make ACL configurable
-	acl := zk.WorldACL(zk.PermRead | zk.PermWrite | zk.PermCreate | zk.PermDelete)
-
+func (c *Client) CreateSequential(path string, data []byte, acl []zk.ACL) (*ZNode, error) {
 	return c.doCreate(path, data, zk.FlagSequence, acl)
 }
 
@@ -202,17 +217,23 @@ func (c *Client) Read(path string) (*ZNode, error) {
 		return nil, fmt.Errorf("failed to read ZNode '%s': %w", path, err)
 	}
 
+	acls, _, err := c.zkConn.GetACL(path)
+	if err != nil {
+		return nil, fmt.Errorf("failed to fetch ACLs for ZNode '%s': %w", path, err)
+	}
+
 	return &ZNode{
 		Path: path,
 		Stat: stat,
 		Data: data,
+		ACL:  acls,
 	}, nil
 }
 
 // Update the ZNode at the given path, under the assumption that it is there.
 //
 // Will return an error if it doesn't already exist.
-func (c *Client) Update(path string, data []byte) (*ZNode, error) {
+func (c *Client) Update(path string, data []byte, acl []zk.ACL) (*ZNode, error) {
 	exists, err := c.Exists(path)
 	if err != nil {
 		return nil, err
@@ -220,6 +241,11 @@ func (c *Client) Update(path string, data []byte) (*ZNode, error) {
 
 	if !exists {
 		return nil, fmt.Errorf("failed to update ZNode '%s': does not exist", path)
+	}
+
+	_, err = c.zkConn.SetACL(path, acl, matchAnyVersion)
+	if err != nil {
+		return nil, fmt.Errorf("failed to update ZNode '%s' ACL: %w", path, err)
 	}
 
 	_, err = c.zkConn.Set(path, data, matchAnyVersion)