+++ internal/provider/provider.go
@@ -26,6 +26,20 @@ func New() (*schema.Provider, error) {
 				Description: "How many seconds a session is considered valid after losing connectivity. " +
 					"More information about ZooKeeper sessions can be found [here](#zookeeper-sessions).",
 			},
+			"username": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Sensitive:   true,
+				DefaultFunc: schema.EnvDefaultFunc(client.EnvZooKeeperUsername, nil),
+				Description: "Username for digest authentication. Can be set via `ZOOKEEPER_USERNAME` environment variable.",
+			},
+			"password": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Sensitive:   true,
+				DefaultFunc: schema.EnvDefaultFunc(client.EnvZooKeeperPassword, nil),
+				Description: "Password for digest authentication. Can be set via `ZOOKEEPER_PASSWORD` environment variable.",
+			},
 		},
 		ResourcesMap: map[string]*schema.Resource{
 			"zookeeper_znode":            resourceZNode(),
@@ -41,9 +55,11 @@ func New() (*schema.Provider, error) {
 func configureProviderContext(_ context.Context, rscData *schema.ResourceData) (interface{}, diag.Diagnostics) {
 	servers := rscData.Get("servers").(string)
 	sessionTimeout := rscData.Get("session_timeout").(int)
+	username := rscData.Get("username").(string)
+	password := rscData.Get("password").(string)
 
 	if servers != "" {
-		c, err := client.NewClient(servers, sessionTimeout)
+		c, err := client.NewClient(servers, sessionTimeout, username, password)
 
 		if err != nil {
 			// Report inability to connect internal Client